Generalize the parser
When we interested a hole (in a `String` position!) in the sanity test module, the plugin panicked with a error about unexpected arguments to `HasField`. This turned out to be `ghc` simply trying all kinds of terms in that hole, the error disappeared with `-fno-show-valid-hole-fits`.

We now generalize the parser in two ways: we recognize variable field labels

```
Record ('(x, a) ': ...) -- @x@ a type variable
```

as well as variable record fields

```
Record (... ': r)       -- @r@ a type variable
```

and we generalized the parsing infrastructure to allow for a more general
predicate to see if we want to attempt to parse a particular constraint at all;
we're using this to ignore `HasField` constraints where the record is not of
the shape `Record r` for some `r`.