JSON support
This is working now, but I'm not sure actually that I like the way it works.
The generic instance always alphabetices the fields, so that show on a record
where the fields are not in order actually gives us a reordered record. Given
that the library does _not_ consider records with different field order to be
equal (merely isomorphiv), this probably should not happen.